Safari-611.1.10/LayoutTests/fast/overflow/scroll-anchor-in-overflow-in-position-fixed-expected.txt

Tests scrolling to an anchor inside overflow:scroll inside position:fixed correctly scrolls the overflow, and not the page

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scroller.scrollTop is 465
PASS scroller.scrollLeft is 0
PASS document.scrollingElement.scrollTop is 800
PASS document.scrollingElement.scrollLeft is 100
PASS successfullyParsed is true

TEST COMPLETE
Anchor is here
